Fix #52 - Во фрагменте раскидал большую функцию по мелким, переименовал немножко репозиторий и интерактор
