+++ app/src/main/java/ru/practicum/android/diploma/di/InteractorModule.kt
@@ -2,9 +2,9 @@ package ru.practicum.android.diploma.di
 
 import org.koin.dsl.module
 import ru.practicum.android.diploma.domain.api.search.SearchInteractor
-import ru.practicum.android.diploma.domain.api.search.details.SharingInteractor
+import ru.practicum.android.diploma.domain.api.details.DetailsInteractor
 import ru.practicum.android.diploma.domain.impl.SearchInteractorImpl
-import ru.practicum.android.diploma.domain.impl.details.SharingInteractorImpl
+import ru.practicum.android.diploma.domain.impl.details.DetailsInteractorImpl
 
 val interactorModule = module {
 
@@ -12,7 +12,7 @@ val interactorModule = module {
         SearchInteractorImpl(get())
     }
 
-    single<SharingInteractor> {
-        SharingInteractorImpl(get())
+    single<DetailsInteractor> {
+        DetailsInteractorImpl(get())
     }
 }
