--- app/src/main/java/ru/practicum/android/diploma/ui/details/DetailsFragment.kt
@@ -1,36 +1,0 @@
-package ru.practicum.android.diploma.ui.details
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.databinding.FragmentDetailsBinding
-
-class DetailsFragment : Fragment() {
-
-    private var _binding: FragmentDetailsBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentDetailsBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.tvVacancy.text = getString(R.string.details_top_nav_header)
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        _binding = null
-    }
-
-}
-
