+++ app/src/main/java/ru/practicum/android/diploma/ui/details/viewmodel/VacancyDescriptionViewModel.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.search.viewmodel
+package ru.practicum.android.diploma.ui.details.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -6,23 +6,18 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
 import ru.practicum.android.diploma.domain.api.search.SearchInteractor
-import ru.practicum.android.diploma.domain.api.search.details.SharingInteractor
+import ru.practicum.android.diploma.domain.api.details.DetailsInteractor
 import ru.practicum.android.diploma.domain.models.VacancyDescription
 
 class VacancyDescriptionViewModel(
     private val interactor: SearchInteractor,
-    private val sharingInteractor: SharingInteractor,
-    vacancyId: String
+    private val detailsInteractor: DetailsInteractor,
 ) : ViewModel() {
 
     private var _vacancyDescriptionState = MutableLiveData<VacancyDescriptionState>()
     val vacancyDescriptionState: LiveData<VacancyDescriptionState> get() = _vacancyDescriptionState
 
-    init {
-        getVacancyDescription(vacancyId)
-    }
-
-    private fun getVacancyDescription(vacancyId: String) {
+    fun getVacancyDescription(vacancyId: String) {
         _vacancyDescriptionState.postValue(VacancyDescriptionState.Loading)
 
         viewModelScope.launch {
@@ -31,7 +26,6 @@ class VacancyDescriptionViewModel(
                     processResult(pair.first, pair.second)
                 }
         }
-
     }
 
     private fun processResult(data: VacancyDescription?, message: String?) {
@@ -40,18 +34,17 @@ class VacancyDescriptionViewModel(
         } else {
             _vacancyDescriptionState.postValue(VacancyDescriptionState.Content(data))
         }
-
     }
 
     fun shareLink(link: String) {
-        sharingInteractor.shareLink(link)
+        detailsInteractor.shareLink(link)
     }
 
     fun openEmail(email: String) {
-        sharingInteractor.openEmail(email)
+        detailsInteractor.openEmail(email)
     }
 
     fun doCall(phone: String) {
-        sharingInteractor.doCall(phone)
+        detailsInteractor.doCall(phone)
     }
 }
