+++ app/src/main/java/ru/practicum/android/diploma/ui/search/fragment/SearchFragment.kt
@@ -19,10 +19,10 @@ import org.koin.androidx.viewmodel.ext.android.viewModel
 import ru.practicum.android.diploma.R
 import ru.practicum.android.diploma.databinding.FragmentSearchBinding
 import ru.practicum.android.diploma.domain.models.Vacancy
+import ru.practicum.android.diploma.ui.details.fragment.VacancyDescriptionFragment
 import ru.practicum.android.diploma.ui.search.adapter.VacancyAdapter
 import ru.practicum.android.diploma.ui.search.viewmodel.SearchState
 import ru.practicum.android.diploma.ui.search.viewmodel.SearchViewModel
-import ru.practicum.android.diploma.util.VACANCY
 import ru.practicum.android.diploma.util.getNumberString
 
 class SearchFragment : Fragment() {
