+++ app/src/main/res/layout/fragment_details.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.search.fragment.VacancyDescriptionFragment">
+    tools:context=".ui.details.fragment.VacancyDescriptionFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/leftBorder"
